# src/main/java/pl/wsb/students/api/model/MovieActorRatingRequest.java
/*
 * Java Course Movie App API
 * Sample Movie App AI
 *
 * OpenAPI spec version: 1.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package pl.wsb.students.api.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import io.swagger.annotations.ApiModelProperty;
import javax.validation.constraints.*;

/**
 * MovieActorRatingRequest
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2020-04-30T07:31:54.867Z")
public class MovieActorRatingRequest   {
  @JsonProperty("movie_id")
  private Integer movieId = null;

  @JsonProperty("actor_id")
  private Integer actorId = null;

  @JsonProperty("rating")
  private Integer rating = null;

  public MovieActorRatingRequest movieId(Integer movieId) {
    this.movieId = movieId;
    return this;
  }

  /**
   * Movie involving actor to rate id
   * @return movieId
   **/
  @JsonProperty("movie_id")
  @ApiModelProperty(required = true, value = "Movie involving actor to rate id")
  @NotNull
  public Integer getMovieId() {
    return movieId;
  }

  public void setMovieId(Integer movieId) {
    this.movieId = movieId;
  }

  public MovieActorRatingRequest actorId(Integer actorId) {
    this.actorId = actorId;
    return this;
  }

  /**
   * Actor to rate
   * @return actorId
   **/
  @JsonProperty("actor_id")
  @ApiModelProperty(required = true, value = "Actor to rate")
  @NotNull
  public Integer getActorId() {
    return actorId;
  }

  public void setActorId(Integer actorId) {
    this.actorId = actorId;
  }

  public MovieActorRatingRequest rating(Integer rating) {
    this.rating = rating;
    return this;
  }

  /**
   * My rate
   * @return rating
   **/
  @JsonProperty("rating")
  @ApiModelProperty(required = true, value = "My rate")
  @NotNull
  public Integer getRating() {
    return rating;
  }

  public void setRating(Integer rating) {
    this.rating = rating;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    MovieActorRatingRequest movieActorRatingRequest = (MovieActorRatingRequest) o;
    return Objects.equals(this.movieId, movieActorRatingRequest.movieId) &&
        Objects.equals(this.actorId, movieActorRatingRequest.actorId) &&
        Objects.equals(this.rating, movieActorRatingRequest.rating);
  }

  @Override
  public int hashCode() {
    return Objects.hash(movieId, actorId, rating);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class MovieActorRatingRequest {\n");
    
    sb.append("    movieId: ").append(toIndentedString(movieId)).append("\n");
    sb.append("    actorId: ").append(toIndentedString(actorId)).append("\n");
    sb.append("    rating: ").append(toIndentedString(rating)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}